refactor: split subcommand classes into separate files
Rather than have the cache command and its subcommands in the same
file, this splits subcommands into dedicated files, and uses
subdirectories to create namespace.
