+++ Corgibytes.Freshli.Cli/Commands/CacheCommand.cs
@@ -1,5 +1,5 @@
 using System.CommandLine;
-using Corgibytes.Freshli.Cli.CommandOptions;
+using Corgibytes.Freshli.Cli.Commands.Cache;
 
 namespace Corgibytes.Freshli.Cli.Commands
 {
@@ -16,25 +16,6 @@ namespace Corgibytes.Freshli.Cli.Commands
         }
     }
 
-    public class CachePrepareCommand : RunnableCommand<CachePrepareCommandOptions>
-    {
-        public CachePrepareCommand()
-            : base("prepare", "Ensures the cache directory exists and contains a valid cache database.")
-        {}
-    }
 
-    public class CacheDestroyCommand : RunnableCommand<CacheDestroyCommandOptions>
-    {
-        public CacheDestroyCommand()
-            : base("destroy", "Deletes the Freshli cache.")
-        {
-            Option<bool> forceOption = new("--force", "Don't prompt to confirm destruction of cache.")
-            {
-                Arity = ArgumentArity.ZeroOrOne
-            };
-
-            AddOption(forceOption);
-        }
-    }
 
 }
