+++ Corgibytes.Freshli.Cli/IoC/FreshliServiceBuilder.cs
@@ -1,6 +1,7 @@
 ﻿using System.CommandLine.Hosting;
 using Corgibytes.Freshli.Cli.CommandOptions;
 using Corgibytes.Freshli.Cli.CommandRunners;
+using Corgibytes.Freshli.Cli.CommandRunners.Cache;
 using Corgibytes.Freshli.Cli.Formatters;
 using Corgibytes.Freshli.Cli.OutputStrategies;
 using Corgibytes.Freshli.Lib;
